Changes to restore Brian's ability to compile
